Ratio fix main UI mouse pointer + scale mouse input by screen resolution

+++ src/menu/ui_shared.h
@@ -399,8 +399,8 @@ typedef struct {
   float			bias;
   int				realTime;
   int				frameTime;
-	int				cursorx;
-	int				cursory;
+	float			cursorx;
+	float			cursory;
 	qboolean	debug;
 
   cachedAssets_t Assets;
